# doc/openapi.yaml
openapi: 3.1.0
info:
  title: Bookly API
  version: 1.0.0
  description: API documentation for Bookly, a personal finance application.
...
paths:
...
  /budgets:
    post:
      servers:
        - url: /v1
      tags:
        - budgets
      summary: Create a new budget
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetRequest"
      responses:
        200:
          description: Budget created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetResponse"
        400:
          $ref: "#/components/responses/ParamError"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        500:
          $ref: "#/components/responses/InternalError"
    get:
      servers:
        - url: /v1
      tags:
        - budgets
      summary: Get all budgets for the current user
      security:
        - bearerAuth: []
      parameters:
        - name: active_only
          in: query
          required: false
          schema:
            type: boolean
          description: Filter to only active budgets
      responses:
        200:
          description: List of budgets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetsResponse"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        500:
          $ref: "#/components/responses/InternalError"
  /budgets/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
        description: Budget ID
    get:
      servers:
        - url: /v1
      tags:
        - budgets
      summary: Get a budget by ID
      security:
        - bearerAuth: []
      responses:
        200:
          description: Budget retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetResponse"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        403:
          $ref: "#/components/responses/ForbiddenError"
        404:
          $ref: "#/components/responses/NotFoundError"
        500:
          $ref: "#/components/responses/InternalError"
    patch:
      servers:
        - url: /v1
      tags:
        - budgets
      summary: Update a budget
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBudgetRequest"
      responses:
        200:
          description: Budget updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetResponse"
        400:
          $ref: "#/components/responses/ParamError"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        403:
          $ref: "#/components/responses/ForbiddenError"
        404:
          $ref: "#/components/responses/NotFoundError"
        500:
          $ref: "#/components/responses/InternalError"
    delete:
      servers:
        - url: /v1
      tags:
        - budgets
      summary: Delete a budget
      security:
        - bearerAuth: []
      responses:
        200:
          description: Budget deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        403:
          $ref: "#/components/responses/ForbiddenError"
        404:
          $ref: "#/components/responses/NotFoundError"
        500:
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization: Bearer {token}"'
    internalTokenAuth:
      type: apiKey
      in: header
      name: INTERNAL-TOKEN
      description: Authentication using a predefined internal token.
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
            example:
              code: 10001
              message: "user not authenticated"
    ForbiddenError:
      description: The user does not have the necessary permissions for a resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
            example:
              code: 10002
              message: "access denied"
    NotFoundError:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
            example:
              code: 10003
              message: "resource not found"
    ParamError:
      description: Invalid parameters provided in the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
            example:
              code: 20001
              message: "invalid parameter: name is required"
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
            example:
              code: 99999
              message: "internal server error"
  schemas:
    EmptyResponse:
      description: Standard empty response wrapper
      type: object
      properties:
        code:
          type: integer
          description: API response code (0 for success)
          example: 0
        data:
          type: null
          nullable: true
          example: null
...
    ErrorResponse:
      description: Standard error response format
      type: object
      properties:
        code:
          type: integer
          description: Application specific error code
        message:
          type: string
          description: A human-readable description of the error
...
    BudgetResponse:
      description: Budget response
      type: object
      properties:
        code:
          type: integer
          description: Response code
          example: 200
        data:
          $ref: "#/components/schemas/jsonBudget"
    BudgetsResponse:
      description: List of budgets response
      type: object
      properties:
        code:
          type: integer
          description: Response code
          example: 200
        data:
          type: array
          items:
            $ref: "#/components/schemas/jsonBudget"
...
    jsonBudget:
      description: Budget data
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: Budget ID
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        user_id:
          type: integer
          format: int32
          description: User ID who owns the budget
        name:
          type: string
          description: Budget name
        category:
          type: string
          description: Budget category
        amount:
          type: string
          description: Budget amount
        period_type:
          type: string
          enum: [monthly, yearly]
          description: Budget period type
        start_date:
          type: string
          format: date-time
          description: Budget start date
        end_date:
          type: string
          format: date-time
          nullable: true
          description: Budget end date
        is_active:
          type: boolean
          description: Whether the budget is active
      required:
        - id
        - created_at
        - updated_at
        - user_id
        - name
        - category
        - amount
        - period_type
        - start_date
        - is_active
    CreateBudgetRequest:
      description: Request to create a budget
      type: object
      properties:
        name:
          type: string
          description: Budget name
          example: "Monthly Food Budget"
        category:
          type: string
          description: Budget category
          example: "Food"
        amount:
          type: string
          description: Budget amount
          example: "500.00"
        period_type:
          type: string
          enum: [monthly, yearly]
          description: Budget period type
          example: "monthly"
        start_date:
          type: string
          format: date-time
          description: Budget start date
          example: "2024-01-01T00:00:00Z"
        end_date:
          type: string
          format: date-time
          nullable: true
          description: Budget end date (optional)
          example: "2024-12-31T23:59:59Z"
      required:
        - name
        - category
        - amount
        - period_type
        - start_date
    UpdateBudgetRequest:
      description: Request to update a budget
      type: object
      properties:
        name:
          type: string
          description: New budget name
          example: "Updated Food Budget"
        category:
          type: string
          description: New budget category
          example: "Groceries"
        amount:
          type: string
          description: New budget amount
          example: "600.00"
        period_type:
          type: string
          enum: [monthly, yearly]
          description: New budget period type
          example: "monthly"
        start_date:
          type: string
          format: date-time
          description: New budget start date
          example: "2024-02-01T00:00:00Z"
        end_date:
          type: string
          format: date-time
          nullable: true
          description: New budget end date
          example: "2024-12-31T23:59:59Z"
        is_active:
          type: boolean
          description: Whether the budget is active
          example: true
...